chore: move form logic to hook
