+++ apps/trading/components/deposit-container/deposit-form.tsx
@@ -1,6 +1,4 @@
-import { FormProvider, useForm } from 'react-hook-form';
-import { zodResolver } from '@hookform/resolvers/zod';
-import { useAccount, useChainId } from 'wagmi';
+import { FormProvider } from 'react-hook-form';
 import { type Squid } from '@0xsquid/sdk';
 import { type Estimate } from '@0xsquid/squid-types';
 
@@ -11,17 +9,12 @@ import { addDecimalsFormatNumber } from '@vegaprotocol/utils';
 
 import { useT } from '../../lib/use-t';
 
-import { useAssetReadContracts } from './use-asset-read-contracts';
-import { useSquidRoute } from './use-squid-route';
 import { SwapInfo } from './swap-info';
-import { type FormFields, type Configs, useFormSchema } from './form-schema';
-import { useNativeBalance } from './use-native-balance';
+import { type Configs } from './form-schema';
 import * as Fields from './fields';
-import BigNumber from 'bignumber.js';
-import { useEvmDeposit } from '../../lib/hooks/use-evm-deposit';
 import { FeedbackDialog, SquidFeedbackDialog } from './feedback-dialog';
-import { useEvmSquidDeposit } from 'apps/trading/lib/hooks/use-evm-squid-deposit';
 import { type TxDeposit, type TxSquidDeposit } from '../../stores/evm';
+import { useDepositForm } from './use-deposit-form';
 
 export const DepositForm = (props: {
   squid: Squid;
@@ -31,138 +24,24 @@ export const DepositForm = (props: {
   onDeposit?: (tx: TxDeposit | TxSquidDeposit) => void;
   minAmount?: string;
 }) => {
-  const { pubKey, pubKeys } = useVegaWallet();
-  const { address } = useAccount();
-
-  const chainId = useChainId();
-  const defaultChain = props.squid.chains.find(
-    (c) => c.chainId === String(chainId)
-  );
-
-  const schema = useFormSchema({ minAmount: props.minAmount });
-  const form = useForm<FormFields>({
-    resolver: zodResolver(schema),
-    defaultValues: {
-      fromAddress: address,
-      fromChain: defaultChain && defaultChain.chainId,
-      fromAsset: '',
-      // fromAddress is just derived from the connected wallet, but including
-      // it as a form field so its included with the zodResolver validation
-      // and shows up as an error if its not set
-      toAsset: props.initialAsset?.id,
-      toPubKey: pubKey,
-      amount: '',
-    },
-  });
-
-  const fields = form.watch();
-
-  const tokens = props.squid.tokens?.filter((t) => {
-    if (!fields.fromChain) return false;
-    if (t.chainId === fields.fromChain) return true;
-    return false;
-  });
-
-  const chain = props.squid.chains.find((c) => c.chainId === fields.fromChain);
-  const toAsset = props.assets?.find((a) => a.id === fields.toAsset);
-  const fromAsset = tokens?.find((t) => t.address === fields.toAsset);
-
-  // Data relating to the selected from asset, like balance on address and allowance
-  // Only relevant if the asset is not a the chains native asset
-  const { data: balanceData, queryKey: balanceDataQueryKey } =
-    useAssetReadContracts({
-      token: fromAsset,
-      configs: props.configs,
-    });
-
-  // Because the read contracts above don't work with native balances we need to
-  // separately get the native token balance
-  const { data: nativeBalance } = useNativeBalance({
-    address,
-    chainId: fields.fromChain,
-  });
-
-  // Check if the from and to asset are the same, if not then we will be using
-  // squid to swap and then deposit via a post hook
-  const isSwap =
-    fields.fromAsset && toAsset
-      ? fields.fromAsset.toLowerCase() !==
-        toAsset?.source.contractAddress.toLowerCase()
-      : undefined;
-
+  const { pubKeys } = useVegaWallet();
   const {
-    data: routeData,
-    error: routeError,
-    isFetching,
-  } = useSquidRoute({
     form,
+    chain,
+    tokens,
     toAsset,
-    enabled: isSwap,
-  });
-
-  const deposit = useEvmDeposit();
-  const squidDeposit = useEvmSquidDeposit();
+    balances,
+    nativeBalance,
+    route,
+    isSwap,
+    squidDeposit,
+    deposit,
+    onSubmit,
+  } = useDepositForm(props);
 
   return (
     <FormProvider {...form}>
-      <form
-        data-testid="deposit-form"
-        onSubmit={form.handleSubmit(async (fields) => {
-          // Get full details of the chosen assets
-          const fromAsset = tokens.find(
-            (t) =>
-              t.address === fields.fromAsset && t.chainId === fields.fromChain
-          );
-          const toAsset = props.assets?.find((a) => a.id === fields.toAsset);
-
-          if (!toAsset || toAsset.source.__typename !== 'ERC20') {
-            throw new Error('no to asset');
-          }
-
-          if (!fromAsset) {
-            throw new Error('no from asset');
-          }
-
-          const isSwapRequired =
-            fromAsset.address.toLowerCase() !==
-            toAsset.source.contractAddress.toLowerCase();
-
-          if (isSwapRequired) {
-            const res = await squidDeposit.write({
-              asset: toAsset,
-              amount: fields.amount,
-              toPubKey: fields.toPubKey,
-              routeData,
-              chainId: Number(fields.fromChain),
-            });
-            props.onDeposit && props.onDeposit(res);
-          } else {
-            // Same asset, no swap required, use normal ethereum bridge
-            // or normal arbitrum bridge to swap
-
-            // Find the matching config for the selected asset
-            const config = props.configs.find(
-              (c) => c.chain_id === toAsset.source.chainId
-            );
-
-            if (!config) {
-              throw new Error(`no bridge for toAsset ${toAsset.id}`);
-            }
-
-            const res = await deposit.write({
-              asset: toAsset,
-              bridgeAddress: config.collateral_bridge_contract
-                .address as `0x${string}`,
-              amount: fields.amount,
-              allowance: (balanceData?.allowance || BigNumber(0)).toString(),
-              toPubKey: fields.toPubKey,
-              chainId: Number(config.chain_id),
-              requiredConfirmations: config.confirmations,
-            });
-            props.onDeposit && props.onDeposit(res);
-          }
-        })}
-      >
+      <form data-testid="deposit-form" onSubmit={onSubmit}>
         <Fields.FromAddress control={form.control} />
         <Fields.FromChain
           control={form.control}
@@ -176,27 +55,26 @@ export const DepositForm = (props: {
         />
         <Fields.Amount
           control={form.control}
-          balanceOf={balanceData?.balanceOf}
-          nativeBalanceOf={nativeBalance}
+          balanceOf={balances.data?.balanceOf}
+          nativeBalanceOf={nativeBalance.data}
         />
         <Fields.ToPubKey control={form.control} pubKeys={pubKeys} />
         <Fields.ToAsset
           control={form.control}
           assets={props.assets}
           toAsset={toAsset}
-          queryKey={balanceDataQueryKey}
-          route={routeData?.route}
+          queryKey={balances.queryKey}
+          route={route.data?.route}
         />
         {isSwap && (
           <div className="mb-4">
-            <SwapInfo route={routeData?.route} error={routeError} />
+            <SwapInfo route={route.data?.route} error={route.error} />
           </div>
         )}
         <SubmitButton
           isSwap={isSwap}
-          isFetchingRoute={isFetching}
-          // isExecuting={squidDeposit.data.status !== 'idle'}
-          estimate={routeData?.route.estimate}
+          isFetchingRoute={route.isFetching}
+          estimate={route.data?.route.estimate}
         />
         {!isSwap && (
           <FeedbackDialog data={deposit.data} onChange={deposit.reset} />
